Text Wrapping - Minor tweaks and refactoring.
