+++ MonoDragons.Core/UserInterface/UI.cs
@@ -2,6 +2,7 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MonoDragons.Core.Memory;
+using MonoDragons.Core.Text;
 
 namespace MonoDragons.Core.UserInterface
 {
@@ -58,7 +59,7 @@ namespace MonoDragons.Core.UserInterface
 
         public static void DrawTextCentered(string text, Rectangle area, Color color)
         {
-            var wrapped = new Behaviors.WrappingText(() => DefaultFont.Font, () => area.Width).Wrap(text);
+            var wrapped = new WrappingText(() => DefaultFont.Font, () => area.Width).Wrap(text);
             var size = DefaultFont.Font.MeasureString(wrapped);
             DrawText(wrapped, GetCenteredPosition(area, size), color);
         }
@@ -66,9 +67,9 @@ namespace MonoDragons.Core.UserInterface
         public static void DrawTextCentered(string text, Rectangle area, Color color, string font)
         {
             var spriteFont = Resources.Load<SpriteFont>(font);
-            var wrapped = new Behaviors.WrappingText(() => spriteFont, () => area.Width).Wrap(text);
-            var size = spriteFont.MeasureString(text);
-            DrawText(text, GetCenteredPosition(area, size), color, font);
+            var wrapped = new WrappingText(() => spriteFont, () => area.Width).Wrap(text);
+            var size = spriteFont.MeasureString(wrapped);
+            DrawText(wrapped, GetCenteredPosition(area, size), color, font);
         }
 
         private static Vector2 GetCenteredPosition(Rectangle area, Vector2 size)
